Fix: `~=` doing “pattern-contains-value”
The pattern-matching operator (`~=`) was strangely not doing the “contains” operation that one would expect for pattern matching— rather, it was doing an “intersection” operation and returning true for any non-zero intersection.  The end result was that the `value` didn't have to fit within the bits of the `pattern` mask; as long as a single bit of the `value` matched it was considered okay.

Additionally, what makes me believe this wasn't the intended behavior is that all of the implementations were symmetric— for any `pattern` & `value` args, you could swap them and it would produce the same results.  That's not how pattern-matching is meant to work.

* Changed `~=` operator implementations to check that the intersection of the `pattern` & `value` is identical to the `value` itself— essentially a bitwise “contains” operation.
* Changed one of the `testPatternMatchingOperator()` asserts to except failure: `.ugly | .scary ~= .scary | .big`